Add Main entry point
