+++ src/java/magma/Main.java
@@ -1,0 +1,18 @@
+package magma;
+
+import java.nio.file.Path;
+
+public class Main {
+    public static void main(String[] args) {
+        Path javaRoot = Path.of("src/java");
+        Path tsRoot = Path.of("src/node");
+        GenerateDiagram.writeTypeScriptStubs(javaRoot, tsRoot).ifPresent(e -> {
+            e.printStackTrace();
+            System.exit(1);
+        });
+        GenerateDiagram.writeDiagram(Path.of("diagram.puml")).ifPresent(e -> {
+            e.printStackTrace();
+            System.exit(1);
+        });
+    }
+}
